code cleaning
git-svn-id: https://svn.code.sf.net/p/mp3val/subversion@31 f84b9ce1-1e83-45eb-b88c-35cee3ba59f7

+++ trunk/sources/out.cpp
@@ -41,7 +41,7 @@ int WriteToFile(int hFile,char *baseptr,int index,int bytes,int iFileSize) {
 		if(index+bytes>iFileSize) bytes=iFileSize-index;
 	}
 
-	while(iBytesWrittenTotal<(DWORD)bytes) {
+	while(iBytesWrittenTotal<bytes) {
 		if(!CrossAPI_WriteFile(hFile,&baseptr[index+iBytesWrittenTotal],bytes-iBytesWrittenTotal,&iBytesWritten)) {
 			return -1;
 		}
